requirements.txt
# 观尘 (See the Dust) 运行环境依赖

# 核心巡检工具
pyyaml>=6.0

# 可选：如果你需要更丰富的终端 UI 体验
# rich>=12.0
